feat: user notifications view
