+++ src/components/molecules/AnalyticsCard/index.tsx
@@ -123,7 +123,7 @@ const AnalyticsCard = ({
 							>
 								<Typography
 									className={clsx(classes.mainInfoText, colorClass)}
-									variant="body1"
+									variant="h6"
 									data-testid="main-info"
 								>
 									{mainInfo}
