+++ src/layouts/Dashboard/components/Topbar/fixtures.ts
@@ -1,0 +1,1 @@
+export const notificationsUnread = [];
